mime: mail parsing completed, the last thing to do is to provide interfaces

+++ httpserver/HTTPServer/mail_cli/ContentTypes.h
@@ -31,7 +31,7 @@ typedef struct {
 
 
 typedef struct mail_content_disposition {
-	long size;
+	long long size;
 	rstring filename;
 	rstring type;
 	rstring creation_date;
